+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 New in 1.3.0
 ============
 
-* Add command `sys` to execute system commands
+* Colour theme can be customized. Add `setConfig` command.
+
+* Add command `sys` to execute system commands.
 
 New in 1.2.1
 ============